fix(ui5-side-navigation): add parent item to submenu and enhance ACC
According to the interaction design, clicking on the parent item in the Overflow menu
should only open the sub-menu, but should not select the parent item.
Clicking on the parent item in the sub-menu selects it and closes the menu -
this is the only way to select the parent item in the overflow menu (the menu lacks
two-click area functionality for the moment).
This is the reason why we add the parent item again in the sub-menu.

List of changes:
- new behavior for Space/Enter keys - always trigger action (navigation)
and do not expand/collapse
- labels for menus in the overflow area
- dual click item in overflow:
   - new aria description for parent item for dual click area
   - duplicate the parent item as a child item
- updated aria-describedby for dual click item in expanded mode

JIRA: BGSOFUIRODOPI-3443
JIRA: BGSOFUIRODOPI-3515